backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-cors>=4.0.0
werkzeug>=3.0.1
Pillow>=10.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
PyJWT>=2.8.0
